Refactor Mason endpoint commands
